reorganize data module
