--- chemprop/data/v2/data.py
@@ -1,639 +1,0 @@
-import threading
-from collections import OrderedDict
-from typing import Dict, Iterator, List, Optional, Union, Tuple
-
-import numpy as np
-from torch.utils.data import DataLoader, Dataset
-from rdkit import Chem
-
-from chemprop.data.scaler import StandardScaler
-from chemprop.featurizers import get_features_generator
-from chemprop.featurizers import BatchMolGraph, MolGraph
-from chemprop.featurizers import is_explicit_h, is_reaction, is_adding_hs, is_mol
-from chemprop.rdkit import make_mol
-
-# Cache of graph featurizations
-CACHE_GRAPH = True
-SMILES_TO_GRAPH: Dict[str, MolGraph] = {}
-
-
-def cache_graph() -> bool:
-    r"""Returns whether :class:`~chemprop.features.MolGraph`\ s will be cached."""
-    return CACHE_GRAPH
-
-
-def set_cache_graph(cache_graph: bool) -> None:
-    r"""Sets whether :class:`~chemprop.features.MolGraph`\ s will be cached."""
-    global CACHE_GRAPH
-    CACHE_GRAPH = cache_graph
-
-
-def empty_cache():
-    r"""Empties the cache of :class:`~chemprop.features.MolGraph` and RDKit molecules."""
-    SMILES_TO_GRAPH.clear()
-    SMILES_TO_MOL.clear()
-
-
-# Cache of RDKit molecules
-CACHE_MOL = True
-SMILES_TO_MOL: Dict[str, Union[Chem.Mol, Tuple[Chem.Mol, Chem.Mol]]] = {}
-
-
-def cache_mol() -> bool:
-    r"""Returns whether RDKit molecules will be cached."""
-    return CACHE_MOL
-
-
-def set_cache_mol(cache_mol: bool) -> None:
-    r"""Sets whether RDKit molecules will be cached."""
-    global CACHE_MOL
-    CACHE_MOL = cache_mol
-
-
-class MoleculeDatapoint:
-    """A :class:`MoleculeDatapoint` contains a single molecule and its associated features and targets."""
-
-    def __init__(
-        self,
-        smiles: List[str],
-        targets: List[Optional[float]] = None,
-        row: OrderedDict = None,
-        data_weight: float = None,
-        gt_targets: List[bool] = None,
-        lt_targets: List[bool] = None,
-        features: np.ndarray = None,
-        features_generator: List[str] = None,
-        phase_features: List[float] = None,
-        atom_features: np.ndarray = None,
-        atom_descriptors: np.ndarray = None,
-        bond_features: np.ndarray = None,
-        overwrite_default_atom_features: bool = False,
-        overwrite_default_bond_features: bool = False,
-    ):
-        """
-        :param smiles: A list of the SMILES strings for the molecules.
-        :param targets: A list of targets for the molecule (contains None for unknown target values).
-        :param row: The raw CSV row containing the information for this molecule.
-        :param data_weight: Weighting of the datapoint for the loss function.
-        :param gt_targets: Indicates whether the targets are an inequality regression target of the form ">x".
-        :param lt_targets: Indicates whether the targets are an inequality regression target of the form "<x".
-        :param features: A numpy array containing additional features (e.g., Morgan fingerprint).
-        :param features_generator: A list of features generators to use.
-        :param phase_features: A one-hot vector indicating the phase of the data, as used in spectra data.
-        :param atom_descriptors: A numpy array containing additional atom descriptors to featurize the molecule
-        :param bond_features: A numpy array containing additional bond features to featurize the molecule
-        :param overwrite_default_atom_features: Boolean to overwrite default atom features by atom_features
-        :param overwrite_default_bond_features: Boolean to overwrite default bond features by bond_features
-
-        """
-        if features is not None and features_generator is not None:
-            raise ValueError("Cannot provide both loaded features and a features generator.")
-
-        self.smiles = smiles
-        self.targets = targets
-        self.row = row
-        self.features = features
-        self.features_generator = features_generator
-        self.phase_features = phase_features
-        self.atom_descriptors = atom_descriptors
-        self.atom_features = atom_features
-        self.bond_features = bond_features
-        self.overwrite_default_atom_features = overwrite_default_atom_features
-        self.overwrite_default_bond_features = overwrite_default_bond_features
-        self.is_mol_list = [is_mol(s) for s in smiles]
-        self.is_reaction_list = [is_reaction(x) for x in self.is_mol_list]
-        self.is_explicit_h_list = [is_explicit_h(x) for x in self.is_mol_list]
-        self.is_adding_hs_list = [is_adding_hs(x) for x in self.is_mol_list]
-
-        if data_weight is not None:
-            self.data_weight = data_weight
-        if gt_targets is not None:
-            self.gt_targets = gt_targets
-        if lt_targets is not None:
-            self.lt_targets = lt_targets
-
-        # Generate additional features if given a generator
-        if self.features_generator is not None:
-            self.features = []
-
-            for fg in self.features_generator:
-                features_generator = get_features_generator(fg)
-                for m, reaction in zip(self.mol, self.is_reaction_list):
-                    if not reaction:
-                        if m is not None and m.GetNumHeavyAtoms() > 0:
-                            self.features.extend(features_generator(m))
-                        # for H2
-                        elif m is not None and m.GetNumHeavyAtoms() == 0:
-                            # not all features are equally long, so use methane as dummy molecule to determine length
-                            self.features.extend(
-                                np.zeros(len(features_generator(Chem.MolFromSmiles("C"))))
-                            )
-                    else:
-                        if m[0] is not None and m[1] is not None and m[0].GetNumHeavyAtoms() > 0:
-                            self.features.extend(features_generator(m[0]))
-                        elif m[0] is not None and m[1] is not None and m[0].GetNumHeavyAtoms() == 0:
-                            self.features.extend(
-                                np.zeros(len(features_generator(Chem.MolFromSmiles("C"))))
-                            )
-
-            self.features = np.array(self.features)
-
-        # Fix nans in features
-        replace_token = 0
-        if self.features is not None:
-            self.features = np.where(np.isnan(self.features), replace_token, self.features)
-
-        # Fix nans in atom_descriptors
-        if self.atom_descriptors is not None:
-            self.atom_descriptors = np.where(
-                np.isnan(self.atom_descriptors), replace_token, self.atom_descriptors
-            )
-
-        # Fix nans in atom_features
-        if self.atom_features is not None:
-            self.atom_features = np.where(
-                np.isnan(self.atom_features), replace_token, self.atom_features
-            )
-
-        # Fix nans in bond_descriptors
-        if self.bond_features is not None:
-            self.bond_features = np.where(
-                np.isnan(self.bond_features), replace_token, self.bond_features
-            )
-
-        # Save a copy of the raw features and targets to enable different scaling later on
-        self.raw_features, self.raw_targets = self.features, self.targets
-        self.raw_atom_descriptors, self.raw_atom_features, self.raw_bond_features = (
-            self.atom_descriptors,
-            self.atom_features,
-            self.bond_features,
-        )
-
-    @property
-    def mol(self) -> List[Union[Chem.Mol, Tuple[Chem.Mol, Chem.Mol]]]:
-        """Gets the corresponding list of RDKit molecules for the corresponding SMILES list."""
-        mol = make_mols(
-            self.smiles, self.is_reaction_list, self.is_explicit_h_list, self.is_adding_hs_list
-        )
-        if cache_mol():
-            for s, m in zip(self.smiles, mol):
-                SMILES_TO_MOL[s] = m
-
-        return mol
-
-    @property
-    def number_of_molecules(self) -> int:
-        """
-        Gets the number of molecules in the :class:`MoleculeDatapoint`.
-
-        :return: The number of molecules.
-        """
-        return len(self.smiles)
-
-    def set_features(self, features: np.ndarray) -> None:
-        """
-        Sets the features of the molecule.
-
-        :param features: A 1D numpy array of features for the molecule.
-        """
-        self.features = features
-
-    def set_atom_descriptors(self, atom_descriptors: np.ndarray) -> None:
-        """
-        Sets the atom descriptors of the molecule.
-
-        :param atom_descriptors: A 1D numpy array of features for the molecule.
-        """
-        self.atom_descriptors = atom_descriptors
-
-    def set_atom_features(self, atom_features: np.ndarray) -> None:
-        """
-        Sets the atom features of the molecule.
-
-        :param atom_features: A 1D numpy array of features for the molecule.
-        """
-        self.atom_features = atom_features
-
-    def set_bond_features(self, bond_features: np.ndarray) -> None:
-        """
-        Sets the bond features of the molecule.
-
-        :param bond_features: A 1D numpy array of features for the molecule.
-        """
-        self.bond_features = bond_features
-
-    def extend_features(self, features: np.ndarray) -> None:
-        """
-        Extends the features of the molecule.
-
-        :param features: A 1D numpy array of extra features for the molecule.
-        """
-        self.features = (
-            np.append(self.features, features) if self.features is not None else features
-        )
-
-    def num_tasks(self) -> int:
-        """
-        Returns the number of prediction tasks.
-
-        :return: The number of tasks.
-        """
-        return len(self.targets)
-
-    def set_targets(self, targets: List[Optional[float]]):
-        """
-        Sets the targets of a molecule.
-
-        :param targets: A list of floats containing the targets.
-        """
-        self.targets = targets
-
-    def reset_features_and_targets(self) -> None:
-        """Resets the features (atom, bond, and molecule) and targets to their raw values."""
-        self.features, self.targets = self.raw_features, self.raw_targets
-        self.atom_descriptors, self.atom_features, self.bond_features = (
-            self.raw_atom_descriptors,
-            self.raw_atom_features,
-            self.raw_bond_features,
-        )
-
-
-class MolGraphDataset(Dataset):
-    r"""A :class:`MoleculeDataset` contains a list of :class:`MoleculeDatapoint`\ s with access to their attributes."""
-
-    def __init__(self, data: List[MoleculeDatapoint]):
-        r"""
-        :param data: A list of :class:`MoleculeDatapoint`\ s.
-        """
-        self._data = data
-        self._batch_graph = None
-
-    def __len__(self) -> int:
-        """
-        Returns the length of the dataset (i.e., the number of molecules).
-
-        :return: The length of the dataset.
-        """
-        return len(self._data)
-
-    def __getitem__(self, idx: int) -> MolGraph:
-        r"""
-        Gets one or more :class:`MoleculeDatapoint`\ s via an index or slice.
-
-        :param item: An index (int) or a slice object.
-        :return: A :class:`MoleculeDatapoint` if an int is provided or a list of
-            :class:`MoleculeDatapoint`\ s if a slice is provided.
-        """
-        d = self._data[idx]
-        mol_graphs_list = []
-        for m in d.mol:
-            if len(d.smiles) > 1 and (d.atom_features is not None or d.bond_features is not None):
-                raise NotImplementedError(
-                    "Atom descriptors are currently only supported with one molecule "
-                    "per input (i.e., number_of_molecules = 1)."
-                )
-
-            mol_graph = MolGraph(
-                m,
-                d.atom_features,
-                d.bond_features,
-                d.overwrite_default_atom_features,
-                d.overwrite_default_bond_features,
-            )
-            mol_graphs_list.append(mol_graph)
-
-        return MolGraph(self._data[idx])
-
-    def smiles(self, flatten: bool = False) -> Union[List[str], List[List[str]]]:
-        """
-        Returns a list containing the SMILES list associated with each :class:`MoleculeDatapoint`.
-
-        :param flatten: Whether to flatten the returned SMILES to a list instead of a list of lists.
-        :return: A list of SMILES or a list of lists of SMILES, depending on :code:`flatten`.
-        """
-        if flatten:
-            return [smiles for d in self._data for smiles in d.smiles]
-
-        return [d.smiles for d in self._data]
-
-    def mols(
-        self, flatten: bool = False
-    ) -> Union[
-        List[Chem.Mol],
-        List[List[Chem.Mol]],
-        List[Tuple[Chem.Mol, Chem.Mol]],
-        List[List[Tuple[Chem.Mol, Chem.Mol]]],
-    ]:
-        """
-        Returns a list of the RDKit molecules associated with each :class:`MoleculeDatapoint`.
-
-        :param flatten: Whether to flatten the returned RDKit molecules to a list instead of a list of lists.
-        :return: A list of SMILES or a list of lists of RDKit molecules, depending on :code:`flatten`.
-        """
-        if flatten:
-            return [mol for d in self._data for mol in d.mol]
-
-        return [d.mol for d in self._data]
-
-    @property
-    def number_of_molecules(self) -> int:
-        """
-        Gets the number of molecules in each :class:`MoleculeDatapoint`.
-
-        :return: The number of molecules.
-        """
-        return self._data[0].number_of_molecules if len(self._data) > 0 else None
-
-    def batch_graph(self) -> List[BatchMolGraph]:
-        r"""
-        Constructs a :class:`~chemprop.features.BatchMolGraph` with the graph featurization of all the molecules.
-
-        .. note::
-           The :class:`~chemprop.features.BatchMolGraph` is cached in after the first time it is computed
-           and is simply accessed upon subsequent calls to :meth:`batch_graph`. This means that if the underlying
-           set of :class:`MoleculeDatapoint`\ s changes, then the returned :class:`~chemprop.features.BatchMolGraph`
-           will be incorrect for the underlying data.
-
-        :return: A list of :class:`~chemprop.features.BatchMolGraph` containing the graph featurization of all the
-                 molecules in each :class:`MoleculeDatapoint`.
-        """
-        if self._batch_graph is None:
-            self._batch_graph = []
-
-            mol_graphs = []
-            for d in self._data:
-                mol_graphs_list = []
-                for s, m in zip(d.smiles, d.mol):
-                    if s in SMILES_TO_GRAPH:
-                        mol_graph = SMILES_TO_GRAPH[s]
-                    else:
-                        if len(d.smiles) > 1 and (
-                            d.atom_features is not None or d.bond_features is not None
-                        ):
-                            raise NotImplementedError(
-                                "Atom descriptors are currently only supported with one molecule "
-                                "per input (i.e., number_of_molecules = 1)."
-                            )
-
-                        mol_graph = MolGraph(
-                            m,
-                            d.atom_features,
-                            d.bond_features,
-                            overwrite_default_atom_features=d.overwrite_default_atom_features,
-                            overwrite_default_bond_features=d.overwrite_default_bond_features,
-                        )
-                        if cache_graph():
-                            SMILES_TO_GRAPH[s] = mol_graph
-                    mol_graphs_list.append(mol_graph)
-                mol_graphs.append(mol_graphs_list)
-
-            self._batch_graph = [
-                BatchMolGraph([g[i] for g in mol_graphs]) for i in range(len(mol_graphs[0]))
-            ]
-
-        return self._batch_graph
-
-    def features(self) -> List[np.ndarray]:
-        """
-        Returns the features associated with each molecule (if they exist).
-
-        :return: A list of 1D numpy arrays containing the features for each molecule or None if there are no features.
-        """
-        if len(self._data) == 0 or self._data[0].features is None:
-            return None
-
-        return [d.features for d in self._data]
-
-    def phase_features(self) -> List[np.ndarray]:
-        """
-        Returns the phase features associated with each molecule (if they exist).
-
-        :return: A list of 1D numpy arrays containing the phase features for each molecule or None if there are no features.
-        """
-        if len(self._data) == 0 or self._data[0].phase_features is None:
-            return None
-
-        return [d.phase_features for d in self._data]
-
-    def atom_features(self) -> List[np.ndarray]:
-        """
-        Returns the atom descriptors associated with each molecule (if they exit).
-
-        :return: A list of 2D numpy arrays containing the atom descriptors
-                 for each molecule or None if there are no features.
-        """
-        if len(self._data) == 0 or self._data[0].atom_features is None:
-            return None
-
-        return [d.atom_features for d in self._data]
-
-    def atom_descriptors(self) -> List[np.ndarray]:
-        """
-        Returns the atom descriptors associated with each molecule (if they exit).
-
-        :return: A list of 2D numpy arrays containing the atom descriptors
-                 for each molecule or None if there are no features.
-        """
-        if len(self._data) == 0 or self._data[0].atom_descriptors is None:
-            return None
-
-        return [d.atom_descriptors for d in self._data]
-
-    def bond_features(self) -> List[np.ndarray]:
-        """
-        Returns the bond features associated with each molecule (if they exit).
-
-        :return: A list of 2D numpy arrays containing the bond features
-                 for each molecule or None if there are no features.
-        """
-        if len(self._data) == 0 or self._data[0].bond_features is None:
-            return None
-
-        return [d.bond_features for d in self._data]
-
-    def data_weights(self) -> List[float]:
-        """
-        Returns the loss weighting associated with each datapoint.
-        """
-        if not hasattr(self._data[0], "data_weight"):
-            return [1.0 for d in self._data]
-
-        return [d.data_weight for d in self._data]
-
-    def targets(self) -> List[List[Optional[float]]]:
-        """
-        Returns the targets associated with each molecule.
-
-        :return: A list of lists of floats (or None) containing the targets.
-        """
-        return [d.targets for d in self._data]
-
-    def gt_targets(self) -> List[np.ndarray]:
-        """ """
-        if not hasattr(self._data[0], "gt_targets"):
-            return None
-
-        return [d.gt_targets for d in self._data]
-
-    def lt_targets(self) -> List[np.ndarray]:
-        """ """
-        if not hasattr(self._data[0], "lt_targets"):
-            return None
-
-        return [d.lt_targets for d in self._data]
-
-    def num_tasks(self) -> int:
-        """
-        Returns the number of prediction tasks.
-
-        :return: The number of tasks.
-        """
-        return self._data[0].num_tasks() if len(self._data) > 0 else None
-
-    def features_size(self) -> int:
-        """
-        Returns the size of the additional features vector associated with the molecules.
-
-        :return: The size of the additional features vector.
-        """
-        return (
-            len(self._data[0].features)
-            if len(self._data) > 0 and self._data[0].features is not None
-            else None
-        )
-
-    def atom_descriptors_size(self) -> int:
-        """
-        Returns the size of custom additional atom descriptors vector associated with the molecules.
-
-        :return: The size of the additional atom descriptor vector.
-        """
-        return (
-            len(self._data[0].atom_descriptors[0])
-            if len(self._data) > 0 and self._data[0].atom_descriptors is not None
-            else None
-        )
-
-    def atom_features_size(self) -> int:
-        """
-        Returns the size of custom additional atom features vector associated with the molecules.
-
-        :return: The size of the additional atom feature vector.
-        """
-        return (
-            len(self._data[0].atom_features[0])
-            if len(self._data) > 0 and self._data[0].atom_features is not None
-            else None
-        )
-
-    def bond_features_size(self) -> int:
-        """
-        Returns the size of custom additional bond features vector associated with the molecules.
-
-        :return: The size of the additional bond feature vector.
-        """
-        return (
-            len(self._data[0].bond_features[0])
-            if len(self._data) > 0 and self._data[0].bond_features is not None
-            else None
-        )
-
-    def normalize_features(
-        self,
-        scaler: StandardScaler = None,
-        replace_nan_token: int = 0,
-        scale_atom_descriptors: bool = False,
-        scale_bond_features: bool = False,
-    ) -> StandardScaler:
-        """
-        Normalizes the features of the dataset using a :class:`~chemprop.data.StandardScaler`.
-
-        The :class:`~chemprop.data.StandardScaler` subtracts the mean and divides by the standard deviation
-        for each feature independently.
-
-        If a :class:`~chemprop.data.StandardScaler` is provided, it is used to perform the normalization.
-        Otherwise, a :class:`~chemprop.data.StandardScaler` is first fit to the features in this dataset
-        and is then used to perform the normalization.
-
-        :param scaler: A fitted :class:`~chemprop.data.StandardScaler`. If it is provided it is used,
-                       otherwise a new :class:`~chemprop.data.StandardScaler` is first fitted to this
-                       data and is then used.
-        :param replace_nan_token: A token to use to replace NaN entries in the features.
-        :param scale_atom_descriptors: If the features that need to be scaled are atom features rather than molecule.
-        :param scale_bond_features: If the features that need to be scaled are bond descriptors rather than molecule.
-        :return: A fitted :class:`~chemprop.data.StandardScaler`. If a :class:`~chemprop.data.StandardScaler`
-                 is provided as a parameter, this is the same :class:`~chemprop.data.StandardScaler`. Otherwise,
-                 this is a new :class:`~chemprop.data.StandardScaler` that has been fit on this dataset.
-        """
-        if len(self._data) == 0 or (
-            self._data[0].features is None
-            and not scale_bond_features
-            and not scale_atom_descriptors
-        ):
-            return None
-
-        if scaler is None:
-            if scale_atom_descriptors and not self._data[0].atom_descriptors is None:
-                features = np.vstack([d.raw_atom_descriptors for d in self._data])
-            elif scale_atom_descriptors and not self._data[0].atom_features is None:
-                features = np.vstack([d.raw_atom_features for d in self._data])
-            elif scale_bond_features:
-                features = np.vstack([d.raw_bond_features for d in self._data])
-            else:
-                features = np.vstack([d.raw_features for d in self._data])
-            scaler = StandardScaler(replace_nan_token=replace_nan_token)
-            scaler.fit(features)
-
-        if scale_atom_descriptors and not self._data[0].atom_descriptors is None:
-            for d in self._data:
-                d.set_atom_descriptors(scaler.transform(d.raw_atom_descriptors))
-        elif scale_atom_descriptors and not self._data[0].atom_features is None:
-            for d in self._data:
-                d.set_atom_features(scaler.transform(d.raw_atom_features))
-        elif scale_bond_features:
-            for d in self._data:
-                d.set_bond_features(scaler.transform(d.raw_bond_features))
-        else:
-            for d in self._data:
-                d.set_features(scaler.transform(d.raw_features.reshape(1, -1))[0])
-
-        return scaler
-
-    def normalize_targets(self) -> StandardScaler:
-        """
-        Normalizes the targets of the dataset using a :class:`~chemprop.data.StandardScaler`.
-
-        The :class:`~chemprop.data.StandardScaler` subtracts the mean and divides by the standard deviation
-        for each task independently.
-
-        This should only be used for regression datasets.
-
-        :return: A :class:`~chemprop.data.StandardScaler` fitted to the targets.
-        """
-        targets = [d.raw_targets for d in self._data]
-        scaler = StandardScaler().fit(targets)
-        scaled_targets = scaler.transform(targets).tolist()
-        self.set_targets(scaled_targets)
-
-        return scaler
-
-    def set_targets(self, targets: List[List[Optional[float]]]) -> None:
-        """
-        Sets the targets for each molecule in the dataset. Assumes the targets are aligned with the
-        datapoints.
-
-        :param targets: A list of lists of floats (or None) containing targets for each molecule. 
-            This must be the same length as the underlying dataset.
-        """
-        if not len(self._data) == len(targets):
-            raise ValueError(
-                "number of molecules and targets must be of same length! "
-                f"num molecules: {len(self._data)}, num targets: {len(targets)}"
-            )
-        for i in range(len(self._data)):
-            self._data[i].set_targets(targets[i])
-
-    def reset_features_and_targets(self) -> None:
-        """Resets the features (atom, bond, and molecule) and targets to their raw values."""
-        for d in self._data:
-            d.reset_features_and_targets()
